Fix Club Data
